Fix: Landing page footer and refactoring of stylesheet (#18)
## Description

Please provide a brief summary of the changes you made and why they are
necessary.

Fixed issue of the landing page where the footer would not scale
correctly in different browsers and screen resolutions.
Refactoring of the css stylesheet, elements with identical
characteristics are now grouped together in css file.

## Type of Change

Please delete options that are not relevant:

- [X] Bug fix (non-breaking change which fixes an issue)
- [X] Refactor (non-breaking change for code readability/structure)

## Checklist

- [X] I have performed a self-review of my own code.
- [X] Tests pass locally with my changes.
- [X] I have updated the documentation where necessary.

## How Has This Been Tested?

Please describe the tests that you ran to verify your changes. Provide
instructions so others can reproduce.

- Tested in multiple Browsers
- Tested by publishing landing page for fork

**Test Configuration**:
- OS: Fedora 41
- Docker Version: 28.1.1